convert files to ts
