--- index.js
@@ -1,3 +1,0 @@
-const { generateRandomUsername } = require('./src/generateRandomUserNames');
-
-module.exports = generateRandomUsername;
