+++ package.json
@@ -2,18 +2,27 @@
   "name": "friendly-username-generator",
   "version": "1.2.0",
   "description": "A non-explicit random username generator",
-  "main": "index.js",
+  "main": "./build/index.js",
+  "types": "./build/index.d.ts",
+  "files": [
+    "build/**/*"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/EvanPeterson1324/friendly-username-generator"
   },
   "scripts": {
-    "test": "jest"
+    "build": "tsc",
+    "test": "jest",
+    "prepublishOnly": "npm test",
+    "prepare": "npm run build"
   },
   "author": "Evan Peterson",
   "license": "MIT",
   "devDependencies": {
-    "jest": "^27.2.0"
+    "@tsconfig/node14": "^1.0.1",
+    "jest": "^27.2.0",
+    "typescript": "^4.5.4"
   },
   "jest": {
     "verbose": true
